[refactor] 구조 개선: 공통 타입 및 constants 분리와 TypingGame 반응형 적용
