+++ apps/frontend/src/constants/game/font.ts
@@ -1,0 +1,7 @@
+import type { FontSize } from "@/types";
+
+export const FONT_SIZE_CLASS: Record<FontSize, string> = {
+  small: "text-base",
+  medium: "text-[24px]",
+  large: "text-3xl",
+};
