+++ apps/frontend/src/router/router.tsx
@@ -3,7 +3,7 @@ import { createBrowserRouter } from "react-router-dom";
 import GameLobbyPage from "@/pages/GameLobbyPage";
 import GamePage from "@/pages/GamePage";
 import LoginPage from "@/pages/LoginPage";
-import { PATH } from "../constants/path";
+import { PATH } from "../constants/route";
 
 export const router = createBrowserRouter([
   {
