+++ apps/frontend/src/types/game/index.ts
@@ -1,0 +1,1 @@
+export * from "./typingGame";
