+++ apps/frontend/src/types/game/typing.ts
@@ -1,0 +1,3 @@
+export type FontSize = "small" | "medium" | "large";
+
+export type MatchState = "correct" | "wrong" | "composing" | "untyped" | "cursor";
